Auto retrieve (#946)
* Call /retrieve after updating database every time data field is changed
* Updated demo with new templates in master
* Update also comp_pipeline and comp_tasks tables, when saving study
* list only file-picker in the getBuiltInServices
* Added access privileges to update pipeline and  cleanup computation_handlers
* - Split handlers from API in computational subsystem
* Created mocker for computational subsystem in test_projects
* Fixes test_project_workflow
* Fixes bad merge 205cea5
* retrieveInput calls are now POSTed and include an array of keys to retrieve in the body
