+++ scripts/demo/confirmations-invitations.csv
@@ -2,20 +2,20 @@ code,user_id,action,data,created_at
 AOuAejUGDv34i9QtxYK61V7GZmCE4B,1,INVITATION,"{
 ""guest"": ""inviteed@foo.com"" ,
 ""host"" : ""inviter@osparc.io""
-}",2019-07-08 13:18:16.368438
+}",2019-07-15 18:17:33.934830
 uQhnK20tuXWdleIRhZaBcmrWaIrb2p,1,INVITATION,"{
 ""guest"": ""inviteed@foo.com"" ,
 ""host"" : ""inviter@osparc.io""
-}",2019-07-08 13:18:16.368446
+}",2019-07-15 18:17:33.934839
 weedI0YvR6tMA7XEpaxgJZT2Z8SCUy,1,INVITATION,"{
 ""guest"": ""inviteed@foo.com"" ,
 ""host"" : ""inviter@osparc.io""
-}",2019-07-08 13:18:16.368449
+}",2019-07-15 18:17:33.934843
 Q9m5C98ALYZDr1BjilkaaXWSMKxU21,1,INVITATION,"{
 ""guest"": ""inviteed@foo.com"" ,
 ""host"" : ""inviter@osparc.io""
-}",2019-07-08 13:18:16.368453
+}",2019-07-15 18:17:33.934846
 jvhSQfoAAfin4htKgvvRYi3pkYdPhM,1,INVITATION,"{
 ""guest"": ""inviteed@foo.com"" ,
 ""host"" : ""inviter@osparc.io""
-}",2019-07-08 13:18:16.368459
+}",2019-07-15 18:17:33.934850
