+++ scripts/demo/portal_markdown.md
@@ -1,4 +1,4 @@
-<!-- Generated by create_portal_markdown.py on 2019-07-08 15:15:04.106152 -->
+<!-- Generated by create_portal_markdown.py on 2019-07-15 16:17:33.906421 -->
 # THE PORTAL Emulator
 
 This pages is for testing purposes for issue [#715](https://github.com/ITISFoundation/osparc-simcore/issues/715)
@@ -8,11 +8,11 @@ This pages is for testing purposes for issue [#715](https://github.com/ITISFound
 
 ## studies available @master
 
-- [ISAN2019: 3D Paraview](http://master.osparc.io/study/template-uuid-518d-a25d-8887bcae93f8)
-- [ISAN: MattWard use case](http://master.osparc.io/study/template-uuid-5a9e-9580-c53d92d18803)
-- [ISAN2019 - opencor-py model](http://master.osparc.io/study/154fb4ad-4913-478f-af04-19725db901a7?stimulation_period_secs=1200)
-- [ISAN: 2D Plot](http://master.osparc.io/study/template-uuid-5716-bedd-b409bb021760)
-- [ISAN: UCDavis use case: 0D](http://master.osparc.io/study/template-uuid-5d82-b08d-d39c436ca738)
+- [ISAN2019: 3D Paraview](http://master.osparc.io/study/template-uuid-518d-a25d-8887bcae93f8) 
+- [ISAN: osparc-opencor](http://master.osparc.io/study/194bb264-a717-11e9-9dff-02420aff2767?stimulation_mode=1&stimulation_level=0.5) with stimulation_mode=1 and stimulation_level=0.5 
+- [ISAN: 2D Plot](http://master.osparc.io/study/template-uuid-5716-bedd-b409bb021760) 
+- [ISAN: UCDavis use case: 0D](http://master.osparc.io/study/template-uuid-5d82-b08d-d39c436ca738) 
+- [ISAN: MattWard use case](http://master.osparc.io/study/template-uuid-5a9e-9580-c53d92d18803) 
 
 ## studies available @staging
