+++ services/web/server/src/simcore_service_webserver/projects/projects_handlers.py
@@ -5,6 +5,7 @@ import logging
 from aiohttp import web
 from jsonschema import ValidationError
 
+from ..computation_api import update_pipeline_db
 from ..login.decorators import RQT_USERID_KEY, login_required
 from ..security_api import check_permission
 from .projects_api import validate_project
@@ -21,6 +22,7 @@ log = logging.getLogger(__name__)
 async def create_projects(request: web.Request):
     # pylint: disable=too-many-branches
     await check_permission(request, "project.create")
+    await check_permission(request, "services.pipeline.*") # due to update_pipeline_db
 
     user_id = request[RQT_USERID_KEY]
     db = request.config_dict[APP_PROJECT_DBAPI]
@@ -69,6 +71,9 @@ async def create_projects(request: web.Request):
 
         # update metadata (uuid, timestamps, ownership) and save
         await db.add_project(project, user_id, force_as_template=as_template is not None)
+
+        # Every change in projects workbench needs to be reflected in the pipeline db
+        await update_pipeline_db(request.app, project["uuid"], project["workbench"])
 
     except ValidationError:
         raise web.HTTPBadRequest(reason="Invalid project data")
@@ -155,13 +160,14 @@ async def replace_project(request: web.Request):
 
     :raises web.HTTPNotFound: cannot find project id in repository
     """
+    await check_permission(request, "services.pipeline.*") # due to update_pipeline_db
 
     user_id = request[RQT_USERID_KEY]
     project_uuid = request.match_info.get("project_id")
     new_project = await request.json()
 
-    db = request.config_dict[APP_PROJECT_DBAPI]
 
+    db = request.config_dict[APP_PROJECT_DBAPI]
     await check_permission(request, "project.update | project.workbench.node.inputs.update",
     context={
         'dbapi': db,
@@ -174,6 +180,9 @@ async def replace_project(request: web.Request):
         validate_project(request.app, new_project)
 
         await db.update_user_project(new_project, user_id, project_uuid)
+
+        # Every change in projects workbench needs to be reflected in the pipeline db
+        await update_pipeline_db(request.app, project_uuid, new_project["workbench"])
 
     except ValidationError:
         raise web.HTTPBadRequest
@@ -205,8 +214,11 @@ async def delete_project(request: web.Request):
     db = request.config_dict[APP_PROJECT_DBAPI]
 
     try:
-        # TODO: this should also delete all the dynamic services used by this project when this happens.
+        # TODO: delete all the dynamic services used by this project when this happens.
+        # TODO: delete pipeline db tasks
+
         await db.delete_user_project(user_id, project_uuid)
+
     except ProjectNotFoundError:
         raise web.HTTPNotFound
